changed file db migration

+++ src/main/resources/db/migration/V1__Init_DB.sql
@@ -7,11 +7,26 @@ DROP TABLE IF EXISTS point CASCADE;
 DROP TABLE IF EXISTS attribute CASCADE;
 DROP TABLE IF EXISTS value CASCADE;
 DROP SEQUENCE IF EXISTS hibernate_sequence;
+-- DROP SEQUENCE IF EXISTS point_id_sequence;
+-- DROP SEQUENCE IF EXISTS attribute_id_sequence;
+-- DROP SEQUENCE IF EXISTS value_id_sequence;
 
 CREATE SEQUENCE hibernate_sequence
   AS BIGINT
   START 1
   INCREMENT 1;
+-- CREATE SEQUENCE point_id_sequence
+--   AS BIGINT
+--   START 1
+--   INCREMENT 1;
+-- CREATE SEQUENCE attribute_id_sequence
+--   AS BIGINT
+--   START 1
+--   INCREMENT 1;
+-- CREATE SEQUENCE value_id_sequence
+--   AS BIGINT
+--   START 1
+--   INCREMENT 1;
 
 CREATE TABLE point
 (
@@ -21,23 +36,27 @@ CREATE TABLE point
   latitude   NUMERIC(6, 3),
   PRIMARY KEY (id)
 );
-
 CREATE TABLE attribute
 (
   id   BIGSERIAL,
   name VARCHAR(255),
   PRIMARY KEY (id)
 );
-
 CREATE TABLE value
 (
   id           BIGSERIAL,
-  point_id     int8 NOT NULL,
-  attribute_id int8 NOT NULL,
+  point_id     BIGINT NOT NULL,
+  attribute_id BIGINT NOT NULL,
   value        VARCHAR(255),
   PRIMARY KEY (point_id, attribute_id)
 );
 
+-- ALTER SEQUENCE point_id_sequence
+-- OWNED BY point.id;
+-- ALTER SEQUENCE attribute_id_sequence
+-- OWNED BY attribute.id;
+-- ALTER SEQUENCE value_id_sequence
+-- OWNED BY value.id;
 ALTER TABLE IF EXISTS value
   ADD CONSTRAINT valueToPoint FOREIGN KEY (point_id) REFERENCES point (id);
 ALTER TABLE IF EXISTS value
